refactor: Rename useMultiframeImages
