+++ packages/insight-viewer/src/Viewer/types.ts
@@ -1,6 +1,6 @@
 import { CornerstoneImage } from '../utils/cornerstoneHelper'
 import { ProgressComponent, Viewport, OnViewportChange } from '../types'
-import { SetFrame } from '../hooks/useFrames/useFrame'
+import { SetFrame } from '../hooks/useMultiframeImages/useFrame'
 import { Interaction } from '../hooks/useInteraction/types'
 
 export type ViewerProp = {
